+++ proj/transform.hpp
@@ -4,22 +4,21 @@
 #define BRIG_PROJ_TRANSFORM_HPP
 
 #include <brig/proj/detail/lib.hpp>
-#include <cstdint>
 #include <stdexcept>
 
 namespace brig { namespace proj {
 
-inline void transform(double* point_xy_begin, uint32_t point_xy_count, projPJ in_pj, projPJ out_pj)
+inline void transform(double* point_xy_begin, long point_xy_count, projPJ in_pj, projPJ out_pj)
 {
   if (detail::lib::singleton().p_pj_is_latlong(in_pj))
-    for (uint32_t i(0), count(2*point_xy_count); i < count; ++i)
+    for (long i(0), count(2*point_xy_count); i < count; ++i)
       point_xy_begin[i] *= DEG_TO_RAD;
 
   if (0 != detail::lib::singleton().p_pj_transform(in_pj, out_pj, point_xy_count, 2, point_xy_begin, point_xy_begin + 1, 0))
     throw std::runtime_error("projection error");
 
   if (detail::lib::singleton().p_pj_is_latlong(out_pj))
-    for (uint32_t i(0), count(2*point_xy_count); i < count; ++i)
+    for (long i(0), count(2*point_xy_count); i < count; ++i)
       point_xy_begin[i] *= RAD_TO_DEG;
 }
